Added Virtual Camera (CineMachine)
Inside the MultipurposeCameraRig prefab, I've added the Virtual Camera which can be used to change the zoom (Lens Ortho Size)! Kelly and I decided so far that 7 is a good zoom distance, so the player can see more of the level.

One thing I'll add later on is collision - so that we can make sure the camera doesn't go outside the level; so following the player, but always staying within specified bounds.

If the MultipurposeCameraRig doesn't follow the player when it's in your scene, just select VirtualCamera and drag your 2DCharacter into the "Follow" and "Look At" properties, lmk if it doesn't work 🙂‍↕️